Move 'Multiplayer' object to 'multiplayer' package
