+++ src/com/reco1l/api/ibancho/RoomAPI.kt
@@ -4,7 +4,7 @@ import com.dgsrz.bancho.security.SecurityUtils
 import com.reco1l.api.ibancho.data.*
 import com.reco1l.api.ibancho.data.Room
 import com.reco1l.api.ibancho.data.RoomTeam
-import com.reco1l.osu.Multiplayer
+import com.reco1l.osu.multiplayer.Multiplayer
 import io.socket.client.IO
 import io.socket.client.Socket
 import io.socket.emitter.Emitter.Listener
