+++ src/com/reco1l/api/ibancho/data/RoomMods.kt
@@ -1,7 +1,7 @@
 package com.reco1l.api.ibancho.data
 
 import com.reco1l.osu.conversion.modsToReadable
-import com.reco1l.osu.Multiplayer
+import com.reco1l.osu.multiplayer.Multiplayer
 import ru.nsu.ccfit.zuev.osu.game.mods.GameMod
 import ru.nsu.ccfit.zuev.osu.game.mods.GameMod.*
 import java.util.*
