+++ src/com/reco1l/osu/multiplayer/RoomChat.kt
@@ -24,7 +24,6 @@ import com.edlplan.ui.fragment.BaseFragment
 import com.reco1l.api.ibancho.RoomAPI
 import com.reco1l.api.ibancho.data.RoomPlayer
 import com.reco1l.osu.mainThread
-import com.reco1l.osu.Multiplayer
 import com.reco1l.toolkt.kotlin.async
 import org.anddev.andengine.input.touch.TouchEvent
 import ru.nsu.ccfit.zuev.osu.RGBColor
