+++ src/com/reco1l/osu/multiplayer/RoomOptions.kt
@@ -29,7 +29,6 @@ import com.reco1l.api.ibancho.data.TeamMode
 import com.reco1l.api.ibancho.data.WinCondition
 import com.reco1l.osu.async
 import com.reco1l.osu.mainThread
-import com.reco1l.osu.Multiplayer
 import ru.nsu.ccfit.zuev.osu.Config
 import ru.nsu.ccfit.zuev.osu.MainActivity
 import ru.nsu.ccfit.zuev.osu.ToastLogger
