+++ src/com/reco1l/osu/multiplayer/RoomScene.kt
@@ -14,11 +14,10 @@ import com.reco1l.api.ibancho.data.WinCondition.*
 import com.reco1l.toolkt.kotlin.runSafe
 import com.reco1l.osu.updateThread
 import com.reco1l.osu.mainThread
-import com.reco1l.osu.Multiplayer
-import com.reco1l.osu.Multiplayer.isConnected
-import com.reco1l.osu.Multiplayer.isRoomHost
-import com.reco1l.osu.Multiplayer.player
-import com.reco1l.osu.Multiplayer.room
+import com.reco1l.osu.multiplayer.Multiplayer.isConnected
+import com.reco1l.osu.multiplayer.Multiplayer.isRoomHost
+import com.reco1l.osu.multiplayer.Multiplayer.player
+import com.reco1l.osu.multiplayer.Multiplayer.room
 import com.reco1l.osu.conversion.modsToString
 import com.reco1l.osu.ui.BeatmapButton
 import com.reco1l.osu.ui.ComposedText
