+++ src/com/reco1l/osu/ui/BeatmapButton.kt
@@ -4,7 +4,7 @@ import com.reco1l.api.ibancho.RoomAPI
 import com.reco1l.api.ibancho.data.PlayerStatus.READY
 import com.reco1l.api.ibancho.data.RoomBeatmap
 import com.reco1l.osu.async
-import com.reco1l.osu.Multiplayer
+import com.reco1l.osu.multiplayer.Multiplayer
 import com.reco1l.osu.beatmaplisting.BeatmapDownloader
 import com.reco1l.osu.beatmaplisting.BeatmapListing
 import com.reco1l.osu.multiplayer.RoomScene
