+++ src/ru/nsu/ccfit/zuev/osu/game/GameScene.java
@@ -16,7 +16,7 @@ import com.reco1l.osu.Execution;
 import com.reco1l.osu.graphics.BlankTextureRegion;
 import com.reco1l.osu.graphics.VideoSprite;
 import com.reco1l.osu.ui.InGameLeaderboard;
-import com.reco1l.osu.Multiplayer;
+import com.reco1l.osu.multiplayer.Multiplayer;
 import com.reco1l.osu.multiplayer.RoomScene;
 
 import com.rian.osu.beatmap.Beatmap;
