+++ src/ru/nsu/ccfit/zuev/osu/menu/ModMenu.java
@@ -4,7 +4,7 @@ import com.edlplan.ui.fragment.InGameSettingMenu;
 import com.reco1l.api.ibancho.RoomAPI;
 import com.reco1l.osu.Execution;
 import com.reco1l.osu.conversion.MultiplayerConverter;
-import com.reco1l.osu.Multiplayer;
+import com.reco1l.osu.multiplayer.Multiplayer;
 import com.reco1l.api.ibancho.data.RoomMods;
 import com.reco1l.osu.multiplayer.RoomScene;
