+++ src/ru/nsu/ccfit/zuev/osu/scoring/ScoringScene.java
@@ -2,7 +2,7 @@ package ru.nsu.ccfit.zuev.osu.scoring;
 
 import com.edlplan.framework.utils.functionality.SmartIterator;
 import com.reco1l.osu.Execution;
-import com.reco1l.osu.Multiplayer;
+import com.reco1l.osu.multiplayer.Multiplayer;
 import com.reco1l.osu.multiplayer.RoomScene;
 import com.reco1l.osu.ui.StatisticSelector;
